master-app/requirements.txt
streamlit
google-generativeai
supabase
aiohttp
python-dotenv
